Move visualization methods out of test/.
