+++ metricflow/test/dataflow_plan_to_svg.py
@@ -1,32 +1,10 @@
-import logging
 import os
-from typing import TypeVar
 
-import graphviz
 from _pytest.fixtures import FixtureRequest
 
-from metricflow.dag.mf_dag import DagNode, MetricFlowDag
-from metricflow.object_utils import random_id
+from metricflow.dag.dag_visualization import DagGraphT, _render_via_graphviz
 from metricflow.test.fixtures.setup_fixtures import MetricFlowTestSessionState
 from metricflow.test.plan_utils import snapshot_path_prefix
-
-logger = logging.getLogger(__name__)
-
-
-DagNodeT = TypeVar("DagNodeT", bound=DagNode)
-
-
-def add_nodes_to_digraph(node: DagNodeT, dot: graphviz.Digraph) -> None:
-    """Adds the node (and parent nodes) to the dot for visualization."""
-    for parent_node in node.parent_nodes:
-        add_nodes_to_digraph(parent_node, dot)
-
-    dot.node(name=node.node_id.id_str, label=node.graphviz_label)
-    for parent_node in node.parent_nodes:
-        dot.edge(tail_name=parent_node.node_id.id_str, head_name=node.node_id.id_str)
-
-
-DagGraphT = TypeVar("DagGraphT", bound=MetricFlowDag)
 
 
 def display_graph_if_requested(
@@ -52,23 +30,3 @@ def display_graph_if_requested(
         )
     _render_via_graphviz(dag_graph=dag_graph, file_path_without_svg_suffix=plan_svg_output_path_prefix)
     mf_test_session_state.plans_displayed += 1
-
-
-def display_dag_as_svg(dag_graph: DagGraphT, mf_config_dir: str) -> str:
-    """Create and display the plan as an SVG in the browser.
-
-    Returns the path where the SVG file was created within "mf_config_dir".
-    """
-    svg_dir = os.path.join(mf_config_dir, "generated_svg")
-    random_file_path = os.path.join(svg_dir, f"dag_{random_id()}")
-    _render_via_graphviz(dag_graph=dag_graph, file_path_without_svg_suffix=random_file_path)
-    return random_file_path + ".svg"
-
-
-def _render_via_graphviz(dag_graph: DagGraphT, file_path_without_svg_suffix: str) -> None:
-    dot = graphviz.Digraph(comment=dag_graph.dag_id, node_attr={"shape": "box", "fontname": "Courier"})
-    # Not quite correct if there are shared nodes.
-    for sink_node in dag_graph.sink_nodes:
-        add_nodes_to_digraph(sink_node, dot)
-    dot.format = "svg"
-    dot.render(file_path_without_svg_suffix, view=True, format="svg", cleanup=True)
